requirements.txt
ipython
jupyterlab
notebook
ruff
ultralytics
opencv-python
lapx
PyYAML
matplotlib
-e .
